+++ rm_projectile_motion/include/rm_projectile_motion/projectile_solver_interface.hpp
@@ -11,6 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 #ifndef RM_PROJECTILE_MOTION__PROJECTILE_SOLVER_INTERFACE_HPP_
 #define RM_PROJECTILE_MOTION__PROJECTILE_SOLVER_INTERFACE_HPP_
 
@@ -21,7 +22,7 @@ class ProjectileSolverInterface
 {
 public:
   // return true if transform successfully.
-  virtual bool solve(float target_x, float target_h, float & angle) = 0;
+  virtual bool solve(double target_x, double target_h, double & angle) = 0;
 };
 
 }  // namespace rm_projectile_motion
